+++ internal/provider/executable_busy_unix.go
@@ -1,0 +1,12 @@
+//go:build !windows
+
+package provider
+
+import (
+	"errors"
+	"syscall"
+)
+
+func isExecutableBusyError(err error) bool {
+	return errors.Is(err, syscall.ETXTBSY)
+}